requirements.txt
adjustText==0.7.3
aiohttp==3.6.2
aiomysql==0.0.20
APScheduler==3.6.3
asn1crypto==0.24.0
async-timeout==3.0.1
asyncio==3.4.3
attrs==19.3.0
beautifulsoup4==4.9.0
certifi==2018.8.24
chardet==3.0.4
colour==0.1.5
cryptography==2.6.1
cycler==0.10.0
DateTime==4.3
demjson==2.2.4
discord==1.0.1
discord.py==1.5.1
distro-info==0.21
dnspython==1.16.0
entrypoints==0.3
fast-autocomplete==0.6.0
filprofiler==0.10.0
httplib2==0.11.3
idna==2.6
jtutils==0.0.6
keyring==17.1.1
keyrings.alt==3.1.1
kiwisolver==1.2.0
leven==1.0.4
matplotlib==3.2.1
multidict==4.7.5
mysql-connector-python==8.0.19
nose==1.3.7
numpy==1.18.3
pandas==1.0.3
Pillow==7.1.2
protobuf==3.6.1
pycrypto==2.6.1
pycurl==7.43.0.2
PyGObject==3.30.4
PyMySQL==0.9.2
pyparsing==2.4.7
PySimpleSOAP==1.16.2
python-apt==1.8.4.1
python-csv==0.0.11
python-dateutil==2.8.1
python-debian==0.1.35
python-debianbts==2.8.2
python-Levenshtein==0.12.0
pytz==2020.1
pyxdg==0.25
reportbug===7.5.3-deb10u1
requests==2.21.0
SecretStorage==2.3.1
six==1.12.0
soupsieve==2.0
tzlocal==2.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.24.1
websockets==8.1
xlrd==1.2.0
xmltodict==0.12.0
yarl==1.4.2
zope.interface==5.1.0
